Specialize window opening only to the main window
There was an issue wherein pop-ups/notifications etc from a window would continue to open within the assigned workspace dispite the parent window having moved to a different workspace. This was fixed by specializing the criteria to include both the window class and type. Set window_type = normal for main windows, now floating windows such as popups dont open in the assigned workspace and will follow to the focused workspace.